+++ crates/dioxus-tailwindcss/src/accessibility.rs
@@ -3,3 +3,7 @@ use crate::ext::*;
 // https://tailwindcss.com/docs/screen-readers
 constant!(sr only);
 constant!(not sr only);
+
+// https://tailwindcss.com/docs/forced-color-adjust
+constant!(forced color adjust auto);
+constant!(forced color adjust none);