+++ crates/dioxus-tailwindcss/src/build.rs
@@ -24,3 +24,26 @@ pub fn tailwindcss<P: AsRef<Path> + Clone>(
         Ok(())
     }
 }
+
+pub fn npx_tailwindcss<P: AsRef<Path> + Clone>(
+    path: P,
+    input: &str,
+    output: &str
+) -> Result<(), Box<dyn std::error::Error>> {
+    let mut cmd = Command::new("npx");
+    cmd.current_dir(path);
+    let result = cmd
+        .arg("tailwindcss")
+        .arg("-i").arg(input)
+        .arg("-o").arg(output).output()?;
+    if !result.status.success() {
+        Err(format!("tailwindcss failed: {} {} -> {:?}", input, output, result).into())
+    } else {
+        println!("cargo:warning=dioxus-tailwind::build::npx_tailwindcss: {} {}", input, output);
+        let stdout = String::from_utf8(result.stdout)?;
+        for line in stdout.lines() {
+            println!("cargo:warning=    {:#?}", line);
+        }
+        Ok(())
+    }
+}