+++ crates/dioxus-tailwindcss/src/effects.rs
@@ -6,7 +6,7 @@ constant!(shadow sm);
 constant!(shadow md);
 constant!(shadow lg);
 constant!(shadow xl);
-constant!(shadow 2xl);
+constant!(shadow "2xl");
 constant!(shadow inner);
 constant!(shadow none);
 
@@ -18,6 +18,7 @@ opacities!(opacity);
 
 // https://tailwindcss.com/docs/mix-blend-mode
 blend!(mix blend);
+constant!(mix blend plus darker);
 constant!(mix blend plus lighter);
 
 // https://tailwindcss.com/docs/background-blend-mode