+++ crates/dioxus-tailwindcss/src/lib.rs
@@ -6,8 +6,9 @@ pub use dioxus_class;
 
 pub mod macros;
 
-pub mod responsive;
 pub mod modifier;
+pub mod responsive;
+pub mod dark_mode;
 pub mod layout;
 pub mod flex;
 pub mod spacing;
@@ -32,8 +33,9 @@ pub mod prelude {
     // macros
     pub use crate::colors;
 
-    pub use crate::responsive::*;
     pub use crate::modifier::*;
+    pub use crate::responsive::*;
+    pub use crate::dark_mode::*;
     pub use crate::layout::*;
     pub use crate::flex::*;
     pub use crate::spacing::*;
@@ -54,12 +56,15 @@ pub mod prelude {
 pub mod ext {
     pub use dioxus_class::ext::*;
 
+    pub use crate::special_ending_internal;
+    pub use crate::special_ending;
     pub use crate::any;
     pub use crate::minus;
     pub use crate::colors;
     pub use crate::blend;
     pub use crate::_1_to_6;
     pub use crate::_1_to_12;
+    pub use crate::_1_to_13;
     pub use crate::size_0_to_96;
     pub use crate::fraction_2_to_4;
     pub use crate::fraction_2_to_6;
@@ -69,4 +74,7 @@ pub mod ext {
     pub use crate::_3xs_to_7xl;
     pub use crate::opacities;
     pub use crate::scales;
+    pub use crate::percents;
+    pub use crate::border_radius;
+    pub use crate::border_width;
 }