+++ crates/dioxus-tailwindcss/src/spacing.rs
@@ -4,10 +4,12 @@ use crate::ext::*;
 size_0_to_96!(p);
 size_0_to_96!(px);
 size_0_to_96!(py);
-size_0_to_96!(pl);
-size_0_to_96!(pr);
+size_0_to_96!(ps);
+size_0_to_96!(pe);
 size_0_to_96!(pt);
+size_0_to_96!(pr);
 size_0_to_96!(pb);
+size_0_to_96!(pl);
 
 // https://tailwindcss.com/docs/margin
 size_0_to_96!(m);
@@ -16,19 +18,22 @@ size_0_to_96!(mx);
 constant!(mx auto);
 size_0_to_96!(my);
 constant!(my auto);
-size_0_to_96!(ml);
-constant!(ml auto);
-size_0_to_96!(mr);
-constant!(mr auto);
+size_0_to_96!(ms);
+constant!(ms auto);
+size_0_to_96!(me);
+constant!(me auto);
 size_0_to_96!(mt);
 constant!(mt auto);
+size_0_to_96!(mr);
+constant!(mr auto);
 size_0_to_96!(mb);
 constant!(mb auto);
+size_0_to_96!(ml);
+constant!(ml auto);
 
 // https://tailwindcss.com/docs/space
 size_0_to_96!(space x);
 constant!(space x reverse);
 
-constant!(space y auto);
 size_0_to_96!(space y);
 constant!(space y reverse);